Add note fragment with viewmodel. Create simple read note screen and implement delete action.

+++ app/src/main/java/com/example/cleannotes/event/Event.kt
@@ -5,7 +5,7 @@ import com.example.domain.model.Note
 sealed class Event
 
 object LoadAllNotes: Event()
-data class GetNoteById(val id: Int): Event()
+data class GetNoteById(val id: Long): Event()
 data class CreateNewNote(val title: String, val description: String): Event()
 data class UpdateNote(val note: Note): Event()
 data class DeleteNote(val note: Note): Event()
